backend/pyproject.toml
[project]
name = "neuronotes-fastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
